refactor llm calculator
